A slight change of responsibilities. The Export's DefaultDirector sends the whole record only once ( calling build.part() )- it no longer traverses the record and no longer sends build.part() repeatedly for every subrecord (like build.part(occurrence), build.part(Occurrenve.getHabitat()), build.part(occurrence.getHabitat().getTerritory())). It is up to the Builder to traverse the record by itself.
Advantages: the Builder can decide what to do with the part of the record (Occurrence for instance) best; The Director won't chew it for him.

Changed after a consultation with Jakub a few days ago. It also matches the Import strategy better - the Parser (the opposite of Builder) - is supposed to return a certain part of the record with all subrecords properly linked (Occurrence->Habitat->Territory). The Builder also receives the record with all subrecords properly linked.