The ExportQuery is generated from a list of stored Restrictions.
The reason why this feature must have been added (consider the following scenario): 
The User executes Search to find some records. He wants to export those records. That's why createExportQuery is called. Then, he wants to export only the first page in another format - another ExportQuery must be created. But what if he opens the Search dialog again, but cancels it without entering anything? All data from the dialog would be lost (deleted) and it would not be possible to generate a correct ExportQuery! That's why the Restrictions from the last Search must be stored and used to generate the ExportQuery.