Habitat fixed: the annoying thing (holder object Village referenced as NearestVillage) that caused inconvenience during the automatic processing was fixed by adding "correct" methods getVillage()/setVillage() and public static final String VILLAGE. This string did not replace the NEARESTVILLAGE so as to ensure compatibility with already written code.
XMLBuilder and XMLParser work properly with UTF-8 (caused problems with searching in the database - records weren't reconstructed properly). ImportMng and ExportMng construct the reader/writer more carefully - with proper encoding.

ABCDBuilder and DCBuilder are to be fixed!

Minor bugfix in the imports/DefaultDirector.java (immutable tables are detected properly now).