Log4j now logs the name of the current thread.
DeleteHabitat method moved from AddEdit to DBLayerUtils. It is leveraged by Delete.

A few fixes to ProgressBar and Task so that the modal state works at least a little, doesn't freeze initially and doesn't freeze randomly after several invocations. Updates to the ProgressBar GUI are made in the Event Dispatch Thread now. The thread usage is however not entirely correct yet unfortunately. You have to be careful when using it.

Clear method added to Selection (we need to clear the selection after delete).

Made StatusBarManager abstract and subclassed it by OverviewStatusManager so that others can make some use of the StatusBarManager too. StatusBarManager is an Observer btw.

Hopefully fixed the occasional NullPointerException during occurrence data fetching from the database.