Partially fixed Bug #7686 - Changes the user made in managers are not propagated. Most of the application is now ready for the propagation. Dialogs for Add, Edit and Search are constructed when Plantlore starts. Then after login data are feeded into these dialogs. To be able to do this I had to add new constructors and methods to AutoTextArea, AutoComboBoxNG3, Assistant (also fixed the AutoComboBoxe's missing border).
ManagerBridge is now registered as an Observer of historyModel, wholeHistoryModel and metadataManagerModel and propagates changes to the rest of the application.
Other managers are to be fixed yet.

Another problem occurred related to Bug #7667 - Column choosing doesn't always work, the table column headers didn't change when the order and number of columns were changed in Settings. Fixed that too.