The moment everyone was waiting for has come. The server is here! :)
Login. A little bit modified, still freezes when an JOptionPane.showMessage() is called. This remains to be solved (although I tried to do all changes to the GUI make in the Event Dispatch Thread). Another thing that is yet to be fixed is that ItemAdd/Edit dialog is displayed incorrectly (columns have wrong sizes).

MaingConfig. Uses the new format of connection information (also known as the infamous triplets) - now referred to as DBInfo.

Middleware.DBLF&RMIDBLF. Modified to use the DBInfo instead of bunch of parameters. Much more flexible!

HibernateDBLayer. Uses different information to connect to the database (using the DatabaseSettings) and to perform the login of the plantlore user.

Server. Updated. Uses EDT to make changes to the GUI, L10n, logging, extended server settings and database settings. Dialogs need to be properly tested (the server as well).


It is higly likely you will have to create your DBInfo from scratch (the previous format is longer valid). I advise you to delete "$HOME/plantlore/plantlore.xml" and create DBInfo using the form Add/Edit (accessible via right click in the Select Database dialog). Also please note that the Authentication dialog requires you to insert the "plantlore-login" (krater, *****) instead of the database engine login (sysdba, masterkey).