DefaultCancelAction.java contains a default implementation of the Cancel Action (closing the form). In most forms there is a cancel button. You don't have to specify this Action over and over again anymore. Just type (in the controller): view.myCancelButton.setAction( new DefaultCancelAction(view) );
The ProgressBar is an abstract class. The DefaultProgressBar is a simple subclass that announces the exception and ends the task (because the only abstract method is the exceptionHandler().

There is also another ProgressBar, ProgressBarEx (as in extended), which is a subclass of the ProgressBar and has the "Cancel" button. This is useful when the User is to be presented with a task that may be canceled (such as export or login procedure). The use is extremely simple:

Task t = new MyTask(...);
new DefaultProgressBar( t, view, true );
t.start()

or

Task t = new MyTask(...);
new DefaultProgressBarEx( t, view, true ); // with the "Cancel" button
t.start;