The Export uses the Task/ProgressBar classes now (which lead to a simplification of the code).
The XMLBuilder can handle files of any size now. The other builders are to be refined yet.

The class Template was renamed to Projection (because the name adheres better with the purpose of this class). And so were the XFilter renamed to FileFormat, XTree to ExtendedTree, and XNode to UserTreeNode.