Minor changes in the ExportMng2 & AuthCtrl (several unused parts omitted).
ExportTask2 fixed - the progress of the task is now announced properly.

MInor chanegs in the Login.

DefaultProgressBar(Ex) stops the task before the error is announced (the reason: if a reconnect-allowing exception occurred and reconnect was indeed selected, there were two progress bars which looked confusing).

DefaultReconnectDialog strings fixed (Common.OK ~> Common.Ok).

ProgressBar informs (logger.debug) which task is in the STOPPING or STOPPED mode.

Task allows repeated start of the task (repeated call of the Task.start() method).

AppCoreCtrl.ReconnectAction fixed. It no longer displays the error message twice when reconnect fails. [The problem was caused by a repeated use of the same login task - its original ProgressBar linked via the Observable-Observer mechanism sprang into life with the task restarted. In fact, in that unfixed version, every time the ReconnectAction was called it created another ProgressBar; this might have lead to a situation where three or more ProgressBars were displayed at the same time - as would be the error message.]