ServerManager makes a heavy use of the Task now. The codebase is read from a configuration file now. In order to set the CODEBASE properly, you have to add a tag into the plantlore.server.xml file. This file is located at the same place as your plantlore.xml configuration file and contains the settings required by the server. It should look like this:
<config>
  <server>
    <port>1099</port>
    <connections>20</connections>
    <perip>3</perip>
    <database>
      <engine>postgresql</engine>
      <port>5432</port>
      <parameter></parameter>
      <masteruser>postgres</masteruser>
    </database>
  </server>
  <codebase>c:\Documents and Settings\yaa\workspace\Plantlore\bin/</codebase>
</config>

The <codebase> is the most important tag of them all. There should be the path (or a .jar) to the directory, where the stubs "are" (meaning: the path to the stubs is valid from that directory on, like net.sf.plantlore.server.HibernateDBLayer_Stub.class).

StandardAction is just a helper class to save some troubles with filling the NAME, DESCRIPTION, etc. of a button/menuitem.