ExportMngCtrl2 dialog's parent view fixed.
TableImportMng & Ctrl simplified (use of DefaultProgressBars, etc.)

Server simplified. The two GUIs for creating a new server and connecting to a running server merged together. No MasterUser and MasterPassword required (in compliance with our previous resolution to create accounts on the database engine).

*******************************************
     IMPORTANT!
*******************************************

Changes made to the Server will afflict the HibernateDBLayer as well! You must either create an account on the database engine you are using (with the username and password you wish to use to connect to the database), or add postgres/postgre (or sysdba/masterkey) into the TUSER table, or tweak the HibernateDBLayer.initialize() temporarily!

This problem will be solved the moment the UserManager will behave the way we've discussed (ie. create a database engine account with every "plantlore"-account).