The updated HibernateDBLayer.java
*******************************************
     IMPORTANT!
*******************************************

Make sure, that you have 
a) either created an account on the database engine you are using (with the username and password you wish to use to connect to the database), 
b) or added postgres/postgre (or sysdba/masterkey) into the TUSER table and made use of that account in Login, 
c) or tweaked the HibernateDBLayer.initialize() temporarily (supplying valid values on lines 168 and 169).

In case c) please make sure your local changes don't get back into the repository, thank you.

 * * * Should you fail to follow the advice above, you most certainly will experience problems while trying to connect to the database for the moment! * * *