Obsolete ItemView and ItemCtrl removed from the repository. They are replaced by ItemView2 and ItemCtrl2.
AppCoreCtrl: Fixed DBLayerFactory creation (the codebase should be specified at that time).


Login: Dialogs and Models re-programmed to conform the new (more secure) DBInfo (no masterkey and masterpassword neither stored nor required).
MainConfig: Updated to reflect the new DBInfo and to be able to obtain Codebase from the configuration file.

RMIDBLayerFactory: Sets the codebase properly when the Factory is created for the first time, so that the remote connections don't fail.

tools.RMI: The codebase is set more carefully.

====================================================================
In order to make remote connections work, you have to modify the configuration file "plantlore.xml" to contain the <codebase> tag and set it properly. It should look like this:

<config>
...
  <codebase>c:\Documents and Settings\yaa\workspace\Plantlore\bin/</codebase> 
...
</config>

Make sure that you do not forget the trailing slash (in case the classes are in a directory). The path the tag contains should be the directory from which the fully qualified names such as net.sf.plantlore.server.HibernateDBLayer_Stub.class are valid.

If this setting is lost, it is highly likely it will no longer be possible to create remote connections - Plantlore cannot guess the position of its stubs!
====================================================================