PlantloreConstants can perform the CLASS to TABLE translation (Occurrence.class ~> Table.OCCURRENCE).
TableImport updated. The UnifiedTableParser can recognize the table (into which the data belong) automatically. Parsers contain a special method initialize() that is called in the Task instead of the TableImportMng. This way the User is presented with a ProgressBar during the (possibly long) parsing of the file. That method returns the CLASS of the detected table. The TableImportTask informs the "Environment" that the table has changed.

AppCoreCtrl supplies the TableImportMng with a callback allowing the "Environment" to be updated (meaning: the Add, Edit and Search dialogs will reload the contents of the updated table). Also the windowClosing operation is handled more carefully - logout() is performed.