PlantloreException and DBLayerException allow proper exception wrapping. You should use the new constructors instead of calling DBLayerException.setError() and DBLayerException.translateSQLState(). The constructor is designed to perform such operations automatically where possible and ensures a unified output! In fact, both setError() and translateSQLState() should be private!!
HibernateDBLayer is a beautiful example of sloppy programming. It is a miracle that it actually works! I made some refinements to the following methods:
* executeInsertXXX()
* executeUpdateXXX()
* executeDeleteXXX()
where XXX means every existing version (History, InTransaction, InTransactionHistory, ..).  
Judging by the state of those SIMPLE methods I am afraid even to think about the quality of the rest of someone's code! Sorry Kovo, reading this is the price you have to pay for my rewriting those methods!