The original code was a little bit confusing about the use of `session.delete()` in deleteXXX() methods. Two of the three deleteXXX() methods used session.delete() while the last one used session.update().
I believe that the original code was correct: all deleteXXX() methods should use session.delete() - and those who seek to "delete a record" (i.e. mark it as deleted) should use the same approach as DBLayerUtils.highLevelDelete() (or even better: use that method!). Otherwise it would not be possible to remove some records from the database, would it?