The War with the Logger was won!
The reason was extremely simple! Recently, I have used Logger in the Record - parent of all holder objects. Since logger is not Serializable, it was impossible to pass *any* of its descendants through the RMI, not just User and Right (thx to Jakub who noticed that). This is why the Logger caused that exception. Mystery solved, no magic I'm afraid, just stupidity. Deleting two lines made the trick.

I made a little change to the return call (the User is returned instead of { User, Right }) during my experiments and I'm lazy to undo it, because it works as well (it is because User carries the Right object itself...)

The remote connection should work flawlessly now. HF