DefaultExceptionHandler returned to its initial state.
Projections applies the list of selected columns to a query (addProjections()) a little bit more carefully. In order to process the record properly - namely to recognize, whether the record is marked for export or not - the Record.ID (= "id") must be add to the list of projections and *processed*, too.
The problem was that ID was not included in the list of properties of the record (as it is in fact an internal number of the database that hosts the record), therefore the ID was neither projected nor reconstructed which is why the record got always rejected. The Occurrence.ID is also required so as to be able to find all author-occurrences referring to that occurrence.

Export: Some final tuning of the *code*. Export will use projections for every format that allows the User to specify which columns should be exported, and it will use the record fetching (Hibernate is an object wrapper of a database - it return fully reconstructed records) otherwise. It may speed up a little bit the Export procedure (it is in fact the Import that has much greater need for speed up than Export).