TableImport: Finishing touches to the code + enriched with JavaDoc. Thoroughly tested (all operations behave exactly as described). The question is whether the behaviour is really fine: if the record R is supposed to be replaced by R', but both R and R' are already in the database, then R is deleted. This has its logic, because the following file may exist (schematically):
<add> Z </add> <update> Z -> X </update>

If this file is imported for the first time, only X is in the database. However, if the file is imported again (by accident or so), then both X and Z would be in the database (if we would not have deleted Z), which is wrong because the result should be the same no matter how many times the file is imported. By deleting the original record R we can achieve that.

OccurrenceImport: JavaDoc added, some final tests remain to be performed.