Reconnect is dispatched by the dispatcher now.
*************************************************
System.gc() is called immediately after the OutOfMemoryError occurs in the table import. It is still error prone - if the creation of the DOM4j tree consumes all memory, a call from EDT may occur and it may not have enough memory -> the program ends in a loop that announces OutOfMemoryException. EDT's behaviour....
*************************************************