Login throws more informative exceptions. It is possible to recognize whether the database is not running or the password or the user name is incorrect. The DatabaseLayerException was improved in order to achieve this functionality.
Add dialog can remember default values and restore them when reloaded or asked to. There is another RFE (to delete some parts of the add dialog - it should be done carefully so as not to clash with the mechanism that restores the default values). Edit dialog is not affected by this change.

Habitat tree allows the User to open the Add dialog (popup menu created by Jakub Kotowski was bound to the tree). It seems to work flawlessly.


InterestingFields.java renamed to DefaultValues.java

HibernateDBLayer contains a bug (more in the bug track list on Berlios).